+++ libs/eco-store/cart/feature/src/lib/eco-store-cart-steps/finish/cart-finish.component.ts
@@ -1,0 +1,10 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'eco-cart-finish',
+  imports: [],
+  templateUrl: './cart-finish.component.html',
+  styleUrl: './cart-finish.component.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class CartFinishComponent {}