+++ libs/eco-store/cart/feature/src/lib/eco-store-cart.routes.ts
@@ -44,6 +44,13 @@ export const ecoStoreCartRoutes: Route[] = [
           ),
       },
       {
+        path: ':id',
+        loadComponent: () =>
+          import('./eco-store-cart-steps/finish/cart-finish.component').then(
+            m => m.CartFinishComponent
+          ),
+      },
+      {
         path: '',
         redirectTo: 'resum',
         pathMatch: 'full',