+++ libs/shared/signal-state/data-access-pocketbase/src/lib/pocketbase.features.ts
@@ -1,10 +1,5 @@
-import {
-  updateState,
-  withDevtools,
-  withDevToolsStub,
-  withImmutableState,
-} from '@angular-architects/ngrx-toolkit';
-import { computed, inject, isDevMode, Type } from '@angular/core';
+import { updateState, withImmutableState } from '@angular-architects/ngrx-toolkit';
+import { computed, inject, Type } from '@angular/core';
 import { tapResponse } from '@ngrx/operators';
 import {
   signalStoreFeature,
@@ -58,7 +53,6 @@ export function withPocketBaseListFeature<
   const defaultState = initialGetListState(customInitialState);
 
   return signalStoreFeature(
-    isDevMode() ? withDevtools(featureName) : withDevToolsStub(featureName),
     withImmutableState<PocketBaseGetListState>({
       ...defaultState,
     }),
@@ -170,7 +164,6 @@ export function withPocketBaseGetOneFeature<
       }>(),
       state: type<EntityState<T>>(),
     },
-    isDevMode() ? withDevtools(featureName) : withDevToolsStub(featureName),
     withImmutableState<PocketbaseGetOne<T>>({ selectedItemId: null as IdType<T> | null }),
     withMethods(store => {
       const showNotification = (type: 'SUCCESS' | 'ERROR', message: string): void => {